backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymupdf
python-dotenv
pydantic
jinja2
playwright
google-generativeai
